+++ resources/lang/en/reading.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'menu' => 'Menu',
+    'on_this_page' => 'On this page',
+    'return_to_top' => 'Return to top',
+    'pager' => 'Pager',
+    'prev' => 'Previous page',
+    'next' => 'Next page',
+];